create user routes

+++ controllers/contacts.js
@@ -2,7 +2,7 @@ import expressAsyncHandler from "express-async-handler";
 import contactModel from "../models/contactModel.js";
 
 const handleGetAllContacts = expressAsyncHandler(async (req, res) => {
-   const contacts = await contactModel.find();
+   const contacts = await contactModel.find({ user_id: req.user.id });
    res.status(200).send({
       message: "All Contacts",
       contacts,
@@ -14,7 +14,7 @@ const fetchContactWithID = expressAsyncHandler(async (req, res, next) => {
    const contact = await contactModel.findById(id);
    if (!contact) {
       const error = new Error(`Contact with id: ${id} not found`);
-      error.type = 'NOT_FOUND'; // Custom error type
+      error.type = "NOT_FOUND"; // Custom error type
       return next(error); // Pass the error to the next middleware (error handler)
    }
 
@@ -29,8 +29,13 @@ const updateContactWithID = expressAsyncHandler(async (req, res, next) => {
    const contact = await contactModel.findById(id);
    if (!contact) {
       const error = new Error(`Contact with id: ${id} not found`);
-      error.type = 'NOT_FOUND';
-      return next(error); 
+      error.type = "NOT_FOUND";
+      return next(error);
+   }
+
+   if (contact.user_id.toString() !== req.user.id) {
+      res.status(403);
+      throw new Error("you can only update your contacts");
    }
 
    const updatedContact = await contactModel.findByIdAndUpdate(id, req.body, {
@@ -47,8 +52,13 @@ const deleteContactWithID = expressAsyncHandler(async (req, res, next) => {
    const contact = await contactModel.findById(id);
    if (!contact) {
       const error = new Error(`Contact with id: ${id} not found`);
-      error.type = 'NOT_FOUND';
-      return next(error); 
+      error.type = "NOT_FOUND";
+      return next(error);
+   }
+
+   if (contact.user_id.toString() !== req.user.id) {
+      res.status(403);
+      throw new Error("you can only delete your contacts");
    }
 
    const deletedContact = await contactModel.findByIdAndDelete(id);
@@ -62,11 +72,16 @@ const createContact = expressAsyncHandler(async (req, res, next) => {
    const { name, phone, email } = req.body;
    if (!name || !phone || !email) {
       const error = new Error("All fields are mandatory");
-      error.type = 'VALIDATION_ERROR'; 
-      return next(error); 
+      error.type = "VALIDATION_ERROR";
+      return next(error);
    }
 
-   const contact = await contactModel.create({ name, phone, email });
+   const contact = await contactModel.create({
+      name,
+      phone,
+      email,
+      user_id: req.user.id,
+   });
 
    res.status(201).send({
       message: "Contact created successfully",
